+++ scripts/build.mjs
@@ -26,6 +26,7 @@ const {
 
 let context = null;
 
+/** @type {import("esbuild").BuildOptions} */
 const config = {
     entryPoints: ["src/entry.ts"],
     bundle: true,