+++ src/core/ui/settings/pages/Plugins/models/bunny.ts
@@ -1,30 +1,43 @@
-import { disablePlugin, enablePlugin, getId, getPluginSettingsComponent, isPluginEnabled, pluginSettings } from "@lib/addons/plugins";
+import { PyoncordIcon } from "@core/ui/settings";
+import { disablePlugin, enablePlugin, getPluginSettingsComponent, isPluginEnabled, pluginSettings } from "@lib/addons/plugins";
 import { BunnyPluginManifest } from "@lib/addons/plugins/types";
 import { useObservable } from "@lib/api/storage";
 
-import { UnifiedPluginModel } from "..";
+import { UnifiedPluginModel } from ".";
 
 export default function unifyBunnyPlugin(manifest: BunnyPluginManifest): UnifiedPluginModel {
     return {
         id: manifest.id,
-        name: manifest.name,
-        description: manifest.description,
-        authors: manifest.authors,
-        isEnabled: () => isPluginEnabled(getId(manifest)),
+        name: manifest.display.name,
+        description: manifest.display.description,
+        authors: manifest.display.authors,
+
+        getBadges() {
+            return [
+                { source: { uri: PyoncordIcon } },
+                // { source: findAssetId("CheckmarkLargeBoldIcon")! }
+            ];
+        },
+        isEnabled: () => isPluginEnabled(manifest.id),
         isInstalled: () => manifest.id in pluginSettings,
         usePluginState() {
             useObservable([pluginSettings]);
         },
         toggle(start: boolean) {
-            start
-                ? enablePlugin(getId(manifest), true)
-                : disablePlugin(getId(manifest));
+            try {
+                start
+                    ? enablePlugin(manifest.id, true)
+                    : disablePlugin(manifest.id);
+            } catch (e) {
+                console.error(e);
+                // showToast("Failed to toggle plugin " + e, findAssetId("Small"));
+            }
         },
         resolveSheetComponent() {
             return import("../sheets/PluginInfoActionSheet");
         },
         getPluginSettingsComponent() {
-            return getPluginSettingsComponent(getId(manifest));
+            return getPluginSettingsComponent(manifest.id);
         },
     };
 }